requirements.txt
openai-whisper
torch
torchaudio
transformers
sentence-transformers
faiss-cpu
pdfplumber
mecab-python3
gradio
numpy
pandas
sqlite3
scikit-learn
requests
Pillow
ffmpeg-python
pyaudio
soundfile
librosa